docs(reatom-jsx-gallery): fix gallery panel state imports
